@@ -1,9 +1,6 @@
 package org.nick.learn.algorithm.aaa_binary_search;
 
-import org.junit.After;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.nick.learn.algorithm.search.aaa_binary_search.BinarySearchUtils;
 
 public class BinarySearchUtilsTest {
@@ -17,35 +14,28 @@ public class BinarySearchUtilsTest {
     int result2 = -1;
 
 
-    @Before
-    public void setUp() throws Exception {
-    }
-
-    @After
-    public void tearDown() throws Exception {
-    }
 
     @Test
     public void binarySearchLeftIncludeRightInclude() {
-        Assert.assertEquals(result1, BinarySearchUtils.binarySearchLeftIncludeRightInclude(nums, target1));
-        Assert.assertEquals(result2, BinarySearchUtils.binarySearchLeftIncludeRightInclude(nums, target2));
+        org.junit.jupiter.api.Assertions.assertEquals(result1, BinarySearchUtils.binarySearchLeftIncludeRightInclude(nums, target1));
+        org.junit.jupiter.api.Assertions.assertEquals(result2, BinarySearchUtils.binarySearchLeftIncludeRightInclude(nums, target2));
     }
 
     @Test
     public void binarySearchLeftIncludeRightExclude() {
-        Assert.assertEquals(result1, BinarySearchUtils.binarySearchLeftIncludeRightExclude(nums, target1));
-        Assert.assertEquals(result2, BinarySearchUtils.binarySearchLeftIncludeRightExclude(nums, target2));
+        org.junit.jupiter.api.Assertions.assertEquals(result1, BinarySearchUtils.binarySearchLeftIncludeRightExclude(nums, target1));
+        org.junit.jupiter.api.Assertions.assertEquals(result2, BinarySearchUtils.binarySearchLeftIncludeRightExclude(nums, target2));
     }
 
     @Test
     public void binarySearchLeftExcludeRightInclude() {
-        Assert.assertEquals(result1, BinarySearchUtils.binarySearchLeftExcludeRightInclude(nums, target1));
-        Assert.assertEquals(result2, BinarySearchUtils.binarySearchLeftExcludeRightInclude(nums, target2));
+        org.junit.jupiter.api.Assertions.assertEquals(result1, BinarySearchUtils.binarySearchLeftExcludeRightInclude(nums, target1));
+        org.junit.jupiter.api.Assertions.assertEquals(result2, BinarySearchUtils.binarySearchLeftExcludeRightInclude(nums, target2));
     }
 
     @Test
     public void binarySearchLeftExcludeRightExclude() {
-        Assert.assertEquals(result1, BinarySearchUtils.binarySearchLeftExcludeRightExclude(nums, target1));
-        Assert.assertEquals(result2, BinarySearchUtils.binarySearchLeftExcludeRightExclude(nums, target2));
+        org.junit.jupiter.api.Assertions.assertEquals(result1, BinarySearchUtils.binarySearchLeftExcludeRightExclude(nums, target1));
+        org.junit.jupiter.api.Assertions.assertEquals(result2, BinarySearchUtils.binarySearchLeftExcludeRightExclude(nums, target2));
     }
 }
