@@ -1,9 +1,7 @@
 package org.nick.learn.algorithm.aab_remove_element;
 
-import org.junit.After;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
+
+import org.junit.jupiter.api.Test;
 import org.nick.learn.algorithm.search.aab_remove_element.RemoveElementUtils;
 
 import java.util.Arrays;
@@ -23,35 +21,29 @@ public class RemoveElementUtilsTest {
 //    输入：nums = [0,1,2,2,3,0,4,2], val = 2
 //    输出：5, nums = [0,1,4,0,3]
 
-    @Before
-    public void setUp() throws Exception {
-    }
 
-    @After
-    public void tearDown() throws Exception {
-    }
 
     @Test
     public void removeElement() {
-        Assert.assertEquals(2, RemoveElementUtils.removeElement(nums1, val1));
-        Assert.assertTrue(Arrays.toString(nums1).startsWith("[2, 2"));
+        org.junit.jupiter.api.Assertions.assertEquals(2, RemoveElementUtils.removeElement(nums1, val1));
+        org.junit.jupiter.api.Assertions.assertTrue(Arrays.toString(nums1).startsWith("[2, 2"));
 
 
-        Assert.assertEquals(5, RemoveElementUtils.removeElement(nums2, val2));
-        Assert.assertTrue(Arrays.toString(nums2).startsWith("[0, 1, 3, 0, 4,"));
+        org.junit.jupiter.api.Assertions.assertEquals(5, RemoveElementUtils.removeElement(nums2, val2));
+        org.junit.jupiter.api.Assertions.assertTrue(Arrays.toString(nums2).startsWith("[0, 1, 3, 0, 4,"));
     }
 
     @Test
     public void removeElementDoublePoint() {
-        Assert.assertEquals(2, RemoveElementUtils.removeElementDoublePoint(nums1, val1));
-        Assert.assertTrue(Arrays.toString(nums1).startsWith("[2, 2"));
+        org.junit.jupiter.api.Assertions.assertEquals(2, RemoveElementUtils.removeElementDoublePoint(nums1, val1));
+        org.junit.jupiter.api.Assertions.assertTrue(Arrays.toString(nums1).startsWith("[2, 2"));
 
 
-        Assert.assertEquals(5, RemoveElementUtils.removeElementDoublePoint(nums2, val2));
-        Assert.assertTrue(Arrays.toString(nums2).startsWith("[0, 1, 3, 0, 4,"));
+        org.junit.jupiter.api.Assertions.assertEquals(5, RemoveElementUtils.removeElementDoublePoint(nums2, val2));
+        org.junit.jupiter.api.Assertions.assertTrue(Arrays.toString(nums2).startsWith("[0, 1, 3, 0, 4,"));
 
 
-        Assert.assertEquals(1,
+        org.junit.jupiter.api.Assertions.assertEquals(1,
                 RemoveElementUtils.removeElementDoublePoint(
                         new int[]{1,2,1,1,1,1,1,1,1,1}, 1));
 
