@@ -1,11 +1,13 @@
 package org.nick.learn.algorithm.sort.relearn202411;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.nick.learn.algorithm.utils.ArrayUtils;
 
 import java.util.Arrays;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 
 public class InsertionSortSolutionTest {
 
