@@ -1,48 +1,11 @@
 package org.nick.learn.algorithm.sort;
 
-import org.junit.Assert;
-import org.nick.learn.algorithm.utils.ArrayUtils;
-
-import java.util.Arrays;
 
 public class CountingSort {
 
-    public static void main(String[] args) {
-        int times = 6;
-
-        long startTimeB = System.currentTimeMillis();
-        for (int i = 0; i < times; i++) {
-            int[] array = ArrayUtils.generateIntArray(times);
-
-            System.out.println("array: " + Arrays.toString(array));
-            int countA = ArrayUtils.count(array);
-
-            sort(array, array.length);
-
-            System.out.println("array: " + Arrays.toString(array));
-            int countB = ArrayUtils.count(array);
-            Assert.assertTrue("array is asc.", ArrayUtils.arrayIsAsc(array));
-            Assert.assertEquals(countA, countB);
-            System.out.println();
-        }
-        long endTimeB = System.currentTimeMillis();
-        System.out.println("geek time: " + (endTimeB - startTimeB) + " ms");
-        System.out.println("-----------------------------------------------------------------------------------------");
 
 
-//        int[] arr = {2, 0, 2};
-//        int countA = ArrayUtils.count(arr);
-//        System.out.println("array: " + Arrays.toString(arr));
-//
-//        sort(arr, arr.length);
-//
-//        int countB = ArrayUtils.count(arr);
-//        System.out.println("array: " + Arrays.toString(arr));
-//        Assert.assertEquals("排序前排序后的数据元素总和一致", countA, countB);
-//        Assert.assertTrue("数组元素是升序的 ASC", ArrayUtils.arrayIsAsc(arr));
-    }
-
-    private static void sort(int[] array, int n) {
+    public static void sort(int[] array, int n) {
 
         // 1. 计算最大最小值
         int min = array[0];
