@@ -1,10 +1,6 @@
 package org.nick.learn.leetcode.problem2974;
 
 import lombok.extern.slf4j.Slf4j;
-import org.junit.Assert;
-import org.nick.learn.algorithm.utils.ArrayUtils;
-
-import java.util.Arrays;
 
 /**
  * 第 6 版本，按照 JDK 源码，在元素个数小于 44 时，选择使用插入排序
@@ -22,7 +18,7 @@ public class Solution6 {
         return nums;
     }
 
-    private void insertionSort(int[] nums) {
+    public void insertionSort(int[] nums) {
         if (nums.length <= 1) {
             return;
         }
@@ -62,20 +58,4 @@ public class Solution6 {
 //    }
 
 
-    public static void main(String[] args) {
-        Solution6 solutionInstance = new Solution6();
-
-        int[] nums;
-        for (int i = 0; i < 10000; i++) {
-            nums = ArrayUtils.generateIntArray10();
-            log.info("before sort: {}", Arrays.toString(nums));
-            solutionInstance.insertionSort(nums);
-            log.info("after sort: {}", Arrays.toString(nums));
-            Assert.assertTrue(ArrayUtils.arrayIsAsc(nums));
-            log.info("");
-        }
-
-        nums = new int[]{5, 4, 2, 3};
-        log.info(Arrays.toString(solutionInstance.numberGame(nums)));
-    }
 }
