@@ -1,10 +1,6 @@
 package org.nick.learn.leetcode.problem2974;
 
 import lombok.extern.slf4j.Slf4j;
-import org.junit.Assert;
-import org.nick.learn.algorithm.utils.ArrayUtils;
-
-import java.util.Arrays;
 
 /**
  * 使用插入排序实现力扣 2974. Minimum Number Game
@@ -38,7 +34,7 @@ class Solution6B {
     }
 
 
-    private void insertionSort(int[] nums) {
+    public void insertionSort(int[] nums) {
         if (nums.length <= 1) {
             return;
         }
@@ -58,32 +54,6 @@ class Solution6B {
     }
 
     public static void main(String[] args) {
-        Solution6B solutionInstance = new Solution6B();
-        int[] nums;
-
-        for (int i = 0; i < 1000; i++) {
-            nums = ArrayUtils.generateIntArray(i);
-            log.info("before sort: " + Arrays.toString(nums));
-            int countBeforeSort = ArrayUtils.count(nums);
-
-            solutionInstance.insertionSort(nums);
-            log.info("after sort: " + Arrays.toString(nums));
-            int countAfterSort = ArrayUtils.count(nums);
-
-            Assert.assertTrue("array is asc", ArrayUtils.arrayIsAsc(nums));
-            Assert.assertEquals(countBeforeSort, countAfterSort);
-
-            System.out.println();
-        }
-
-        nums = new int[]{5, 4, 2, 3};
-        solutionInstance.numberGame(nums);
-        Assert.assertArrayEquals(new int[]{3, 2, 5, 4}, nums);
-
-
-        nums = new int[]{2, 5};
-        solutionInstance.numberGame(nums);
-        Assert.assertArrayEquals(new int[]{5, 2}, nums);
 
 
     }
