@@ -1,7 +1,7 @@
 package org.nick.learn.leetcode.problem427;
 
 import lombok.extern.slf4j.Slf4j;
-import org.junit.Assert;
+
 
 @Slf4j
 public class SolutionB {
@@ -38,30 +38,5 @@ public class SolutionB {
         return pre;
     }
 
-    public static void main(String[] args) {
-        SolutionB solution = new SolutionB();
-        int[][] grid;
-        Node root;
 
-        grid = new int[][]{
-                {0, 1},
-                {1, 0}
-        };
-        root = solution.construct(grid);
-        Assert.assertEquals("[0,1,1,0]", root.toString());
-
-        grid = new int[][]{
-                {1, 1, 1, 1, 0, 0, 0, 0},
-                {1, 1, 1, 1, 0, 0, 0, 0},
-                {1, 1, 1, 1, 1, 1, 1, 1},
-                {1, 1, 1, 1, 1, 1, 1, 1},
-                {1, 1, 1, 1, 0, 0, 0, 0},
-                {1, 1, 1, 1, 0, 0, 0, 0},
-                {1, 1, 1, 1, 0, 0, 0, 0},
-                {1, 1, 1, 1, 0, 0, 0, 0}
-        };
-
-        root = solution.construct(grid);
-        Assert.assertEquals("[1,[0,0,1,1],1,0]", root.toString());
-    }
 }
